chore: improved folder structure
