+++ src/components/SnippetComponents/DropdownSnippets.jsx
@@ -1,0 +1,43 @@
+import React, { useState } from "react";
+import Modal from "../Modal";
+import StringToReactComponent from "string-to-react-component";
+import { dropDownSnippets } from "./Snippets/DropDowns";
+
+const DropdownSnippets = () => {
+  const [showModal, setShowModal] = useState(false);
+  const [jsxCode, setJsxCode] = useState("");
+  const [cssCode, setCssCode] = useState("");
+
+  const handleShowModal = (jsx, css) => {
+    setJsxCode(jsx);
+    setCssCode(css);
+    setShowModal(true);
+  };
+
+  return (
+    <div className="grid grid-cols-1 md:grid-cols-3 gap-6">
+      {dropDownSnippets.map((snippet, index) => (
+        <div key={index} className="p-8 bg-white rounded-lg shadow-lg">
+          <h2 className="text-xl font-bold mb-4">{snippet.title}</h2>
+          <StringToReactComponent>{snippet.jsxCode}</StringToReactComponent>
+          <div className="mt-4 flex justify-end">
+            <button
+              className="text-black text-md py-2 px-4 rounded-lg shadow-md"
+              onClick={() => handleShowModal(snippet.jsxCode, snippet.cssCode)}
+            >
+              Show Code
+            </button>
+          </div>
+        </div>
+      ))}
+      <Modal
+        showModal={showModal}
+        onClose={() => setShowModal(false)}
+        jsxCode={jsxCode}
+        cssCode={cssCode}
+      />
+    </div>
+  );
+};
+
+export default DropdownSnippets;
